Implement language server protocol
The language server protocol is largely generated from the
specification
  src/languageserver/3rdparty/specification.md
by the src/languageserver/generate.ts script.

Objects used in the protocol are the POD like C++ objects defined in
qlanguageserverspectypes_p.h.

The QLanguageServerProtocol objects handles the language server
protocol both to send and receive (ask the client and reply to the
client).
To ask the client you use the requestXX or notifyXX methods.
To reply to client request you have to register your handlers via
registerXXRequestHandler, notifications can be handled via
registerXXNotificationHandler signals.
QLspNotifySignals can be used to emit signals for all notifications.

The method themselves are implemented in qlanguageservergen*
files which are generated form the specification.
You have to provide a function to send the data that the protocol
generates to the constructor, and you have to feed the data you
receive to it via the receivedData method.

QLanguageServer is a class that uses the QLanguageServerProtocol to
provide a server implementation.
It handles the lifecycle management, and can be extended via
QLanguageServerModule subclasses.

Limitations: the client use case (Creator) the handling of partial
results could be improved. A clean solution should handle the progress
notification, do some extra tracking and give the partial results back
to the call that did the request.

Change-Id: I087b865e3aad65fbc300d3429c1b408d4e26e299
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>