Add typed json support
typed json support allows one to use C++ POD like objects
and convert them from/to json. This allows support for jsonrpc
with typed C++ objects.
Potentially we could serialize/deserialize directly, and even keep
utf-8 encoding, but to be cleaner we serialize through QJson, and
implement it strictly on the top of QJsonRpcProtocol.

The serialization needs a method walk to be implemented for the
objects to serialize, the implementation is tailored to jsonrpc, it
could be extended (for example to allow non templatized walk), but
this was not done to keep it as simple and small as possible, while
still being flexible enough.

The idea of TypedJsonRpc is slightly different than QJsonRpcProtocol:

Typed handlers are simply functions that receive a method (as QByteArray
that contains an Utf-8 string), the request or notification object (as
a typed C++ object), and either a reference to the typed protocol (for
notifications, to communicate), or a response type built on the top of
the TypedResponse object.

TypedResponse is a move only response object that can be transferred
to another thread. This way multithreading in the hand of the handler
implementor: the handle should *always* return control quickly, if the
response takes a long time then it is up to him to execute it in
another thread, and send the response later.
The move only response ensures that a response (and only one) is
send (send an error in the destructor if no response was sent).
The response can be derived from TypedResponse to accept only specific
typed objects as response.
This way the handler function can specify both the expected input and
output (return type) and thus the protocol.

It is possible to add actions to the reponse that are executed after
the response has been sent, and also add a preprocessor function to
log/trace, or handle non conformant requests.

Change-Id: Ie74e2310b21d8e3c6cc858ec84fc4c118a06632c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>