Remove threading from JsonRpc protocol
Using std::thread for the threading locks the user into a specific kind
of parallelism. They might not want to use that one.

The history of this threading feature requires some explanation. The
first client of this API is the QML language server. The QML language
server runs on the host of a host/target pair, not on the target
machine. Only this way it can interact with an IDE in a sane way. You
don't want to deploy a language server to an embedded device or a mobile
phone. Host tools had to be built with the Qt bootstrap library in Qt5.
The bootstrap library neither provides an event loop nor a threading
abstraction.

Considering all this, the original plan was to use some nonblocking
container as a poor man's event queue and hide the threading inside the
protocol. This may or may not have been a good idea, but we can clearly
state two things about it today:

1. The threading is not really hidden. This is exemplified by the need
   to add mutexes on higher layers.
2. The whole thing is unnecessary in Qt6. In Qt6, we can build host
   tools using the full power of QtCore, which gives us event loops and
   QThread.

Change-Id: I61498de63cf64dd889b7f2471a12a5033038661b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>