languageserver: set default value of variant containing null to null
the specification contains several types that are not optional, but
are a variant of several options, including null.
In these variants null is typically the last option listed, but when
initializing a default value null is the correct default value, not a
default initialized type of the first option.
So detect these types and initialize them with nullptr instead of {}.

Change-Id: I8e28b3aa0a4fdad8f04f8803c027c4c202b41e84
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>