Implement "dry run" mode for QHttpMessageStreamParser
Use the QHttpMessageStreamParser without buffers such that its FSM can
be used, e.g., when reading blocks of data from stdin (and knowing when
to stop and when to wait for more data). When the UNBUFFERED mode is set,
the handlers for message bodies returns empty bodies and the message
bodies are not saved into the QHttpMessageStreamParser.

Task-number: QTBUG-104552
Change-Id: I446c0955ef641097d2c7e102df3bf4c3f689140e
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@qt.io>