qtlanguageserver: Add security header
Everything but the (insignificant) global module header is marked as
significant; all the actually critical parsing in C++ code is happening
at the JSON-RPC layer.
Note that most files are auto-generated, they will  pick the security
line from the header of the generate.ts script for now. If there's ever
a deviation, that would need to be adjusted.
The generate.ts scipt is also marked as (only) significant, as it is
expected to be used only with trusted input, and is not actually run
unless we update the LSP version.

Pick-to: 6.11 6.10 6.9 6.8
Fixes: QTBUG-135760
QUIP: 23
Change-Id: Ie328e8fc07434e2ab6c6454c898c17b4a45bb959
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>