lsp: fix structs for progress notification
It seems that the LSP specification has a templated ProgressParam
struct that we don't generate. We can't make it a templated C++ type
without making all the methods using it templated (like notifyProgress()
or the method that allows to register a callback on the progress
notification). We also can't make it a QVariant or std::any without
breaking the deserialization mechanism (its not clear how to parse JSON
into a QVariant or std::any). Also, its the only templated in the
specification, even in the future one (3.18).

Generate some C++ code for ProgressParams (it was currently special
cased and ignored) and all methods accepting ProgressParams. Replace
the templated member `T value` of ProgressParams with a std::variant
containing all structs that the specification claims can be used with,
and ignore the "generic" part.

This will allow to show some progressbar to the user for QTBUG-142352,
so that the user knows that something is busy in the background.

Also set the 'kind' member of the WorkDoneProgressBegin,
WorkDoneProgressReport and WorkDoneProgressEnd structs: its a hard-coded
constant in the LSP protocol and it shouldn't be changed by the user so
give it a meaningful default value. 'kind' sadly can't be constexpr because
the QTypedJson magic does not support const fields.

Task-number: QTBUG-142352
Change-Id: If4e8fbecefa128d7f3b5dde444864cdc14d4a6e6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>